vestigekv: quantize scan-operand storage (side bf16, csk fp16), fp32 ieee arithmetic
Storage halves tier-2 scan traffic per archive row (516 -> 260 bytes: side
64x2B + csk 64x2B + rho 4B); all dot products still promote to fp32 in
registers with input_precision=ieee, so the arithmetic path is unchanged.
Conformal zp recalibrates on the quantized score path (quantize-then-
calibrate), keeping the certificate valid by construction. kept_rows bf16
is bit-exact against the bf16 pool source. csk carries an fp16-range
assert (amax < 6e4). Live-archive caches (_side_all/_csk_all) stored in
the same dtypes so refresh_membership cannot regress storage to fp32.

Fire-set stability vs fp32 reference (S=256k synthetic, 256 queries):
Jaccard mean 0.9957, exact-top8 coverage 0.9866 vs 0.9863 (parity), zp
4.5382 -> 4.5414 (recalibration absorbing quantization noise). 67/67
unit+kernel tests pass; batched-vs-per-pair paths remain bit-identical.

+++ python/sglang/srt/layers/attention/vestigekv/batched_step.py
@@ -48,16 +48,17 @@ def _scan_batched_kernel(
     d = tl.arange(0, DD)
     r = tl.arange(0, R)
     h = tl.arange(0, H)
+    # quantized storage, fp32 ieee arithmetic (see scan_kernel)
     s = tl.load(
         side_ptr + p * Amax * DD + offs[:, None] * DD + d[None, :],
         mask=m[:, None],
         other=0.0,
-    )
+    ).to(tl.float32)
     c = tl.load(
         csk_ptr + p * Amax * R + offs[:, None] * R + r[None, :],
         mask=m[:, None],
         other=0.0,
-    )
+    ).to(tl.float32)
     rh = tl.load(rho_ptr + p * Amax + offs, mask=m, other=0.0)
     qs = tl.load(qside_t_ptr + p * DD * H + d[:, None] * H + h[None, :])
     qk = tl.load(qsk_t_ptr + p * R * H + r[:, None] * H + h[None, :])
@@ -92,10 +93,10 @@ class BatchedScanPack:
         NKm = int(max(t.kept_rows.shape[0] for t in tiers) * self.HEADROOM)
         Am = int(max(t.side.shape[0] for t in tiers) * self.HEADROOM)
         r = tiers[0].r
-        self.kr = torch.zeros(P, NKm, D.LATENT_DIM, device=dev)
+        self.kr = torch.zeros(P, NKm, D.LATENT_DIM, device=dev, dtype=torch.bfloat16)
         self.v = torch.zeros(P, r, D.KV_LORA_RANK, device=dev)
-        self.side = torch.zeros(P, Am, D.SIDECAR_DIM, device=dev)
-        self.csk = torch.zeros(P, Am, r, device=dev)
+        self.side = torch.zeros(P, Am, D.SIDECAR_DIM, device=dev, dtype=torch.bfloat16)
+        self.csk = torch.zeros(P, Am, r, device=dev, dtype=torch.float16)
         self.rho = torch.zeros(P, Am, device=dev)
         self.arch = torch.zeros(P, Am, dtype=torch.int64, device=dev)
         self.a_len = torch.zeros(P, dtype=torch.int64, device=dev)
@@ -169,7 +170,7 @@ class BatchedScanPack:
         sc = self.scale
         W = self.fetch_buf.shape[-1]
         qe = self.qbuf[self.li, self.slot].float()  # [P, H, 576], one gather+cast
-        skept = torch.bmm(qe, self.kr.transpose(1, 2))
+        skept = torch.bmm(qe.to(torch.bfloat16), self.kr.transpose(1, 2)).float()
         skept.mul_(sc)
         skept.masked_fill_(self.nk_mask, float("-inf"))
         max1 = skept.max(-1).values
